I can set the datepicker date and time properties directly from golang using the following golang: time.Now().String() The exact string golang passes to the ubuntu datepicker "date" property is: 2014-04-27 11:34:17.425730025 -0400 EDT
Essentially we are creating a QML Date using a string.  Unfortunately, when using the datepicker to simply select time, we are forced to create an entire date object and from there we set the hours/minutes/seconds within the string before we pass it back to qml.  It would have been more effective to have direct access to the object, but this workaround will suffice and it removes the necessity to have a middleman qml function to create a date and set it from there.